LayoutTests/crypto/subtle/aes-ctr-import-jwk-key-length-192.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test importing a JWK AES-CTR key with legnth 192");

jsTestIsAsync = true;

var jwkKey = {
    kty: "oct",
    k: "A72FD48989ED7E92A3B3A080F74FA80B",
    alg: "A192CTR",
};
var extractable = true;

debug("Importing a key...");
crypto.subtle.importKey("jwk", jwkKey, "aes-ctr", extractable, ["encrypt", "decrypt", "wrapKey", "unwrapKey"]).then(function(result) {
    key = result;

    shouldBe("key.type", "'secret'");
    shouldBe("key.extractable", "true");
    shouldBe("key.algorithm.name", "'AES-CTR'");
    shouldBe("key.algorithm.length", "192");
    shouldBe("key.usages", "['decrypt', 'encrypt', 'unwrapKey', 'wrapKey']");

    finishJSTest();
});

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
